+++ src/Layers/xrRender/r__dsgraph_render.cpp
@@ -146,7 +146,7 @@ void		sort_tlist_nrm
  BOOL	bSSA
  )
 {
-	int amount			= textures.begin()->key->size();
+	int amount			= (int)textures.begin()->key->size();
 	if (bSSA)	
 	{
 		if (amount<=1)
@@ -192,7 +192,7 @@ void		sort_tlist_mat
  BOOL	bSSA
  )
 {
-	int amount			= textures.begin()->key->size();
+	int amount			= (int)textures.begin()->key->size();
 	if (bSSA)	
 	{
 		if (amount<=1)
