+++ src/Layers/xrRenderDX10/3DFluid/dx103DFluidManager.cpp
@@ -758,7 +758,7 @@ void dx103DFluidManager::UpdateObstacles( const dx103DFluidData &FluidData, floa
 #ifdef	DEBUG
 void dx103DFluidManager::RegisterFluidData(dx103DFluidData* pData, const xr_string &SectionName)
 {
-	int iDataNum = m_lstFluidData.size();
+	int iDataNum = (int)m_lstFluidData.size();
 
 	int i;
 
@@ -781,7 +781,7 @@ void dx103DFluidManager::RegisterFluidData(dx103DFluidData* pData, const xr_stri
 
 void dx103DFluidManager::DeregisterFluidData(dx103DFluidData* pData)
 {
-	int iDataNum = m_lstFluidData.size();
+	int iDataNum = (int)m_lstFluidData.size();
 
 	int i;
 
@@ -806,7 +806,7 @@ void dx103DFluidManager::DeregisterFluidData(dx103DFluidData* pData)
 
 void dx103DFluidManager::UpdateProfiles()
 {
-	int iDataNum = m_lstFluidData.size();
+	int iDataNum = (int)m_lstFluidData.size();
 
 	int i;
