+++ src/Layers/xrRenderDX10/dx10ResourceManager_Resources.cpp
@@ -509,8 +509,8 @@ void	CResourceManager::_DeleteGS			(const SGS* gs)
 static BOOL	dcl_equal			(D3DVERTEXELEMENT9* a, D3DVERTEXELEMENT9* b)
 {
 	// check sizes
-	u32 a_size = GetDeclLength(a);
-	u32 b_size = GetDeclLength(b);
+	u32 a_size = (u32)GetDeclLength(a);
+	u32 b_size = (u32)GetDeclLength(b);
 	if (a_size!=b_size)	return FALSE;
 	return 0==memcmp	(a,b,a_size*sizeof(D3DVERTEXELEMENT9));
 }
@@ -526,7 +526,7 @@ SDeclaration*	CResourceManager::_CreateDecl	(D3DVERTEXELEMENT9* dcl)
 
 	// Create _new
 	SDeclaration* D			= xr_new<SDeclaration>();
-	u32 dcl_size = GetDeclLength(dcl) + 1;
+	u32 dcl_size = (u32)GetDeclLength(dcl) + 1;
 	//	Don't need it for DirectX 10 here
 	//CHK_DX					(HW.pDevice->CreateVertexDeclaration(dcl,&D->dcl));
 	D->dcl_code.assign		(dcl,dcl+dcl_size);
@@ -644,7 +644,7 @@ SGeometry*	CResourceManager::CreateGeom	(D3DVERTEXELEMENT9* decl, ID3DVertexBuff
 	R_ASSERT			(decl && vb);
 
 	SDeclaration* dcl	= _CreateDecl			(decl);
-	u32 vb_stride = ComputeVertexSize(decl, 0);
+	u32 vb_stride = (u32)ComputeVertexSize(decl, 0);
 
 	// ***** first pass - search already loaded shader
 	for (u32 it=0; it<v_geoms.size(); it++)
