+++ src/xrEngine/IGame_Persistent.cpp
@@ -174,9 +174,9 @@ void IGame_Persistent::OnFrame		()
 		Environment().OnFrame	();
 
 
-	Device.Statistic->Particles_starting= ps_needtoplay.size	();
-	Device.Statistic->Particles_active	= ps_active.size		();
-	Device.Statistic->Particles_destroy	= ps_destroy.size		();
+	Device.Statistic->Particles_starting= (u32)ps_needtoplay.size	();
+	Device.Statistic->Particles_active	= (u32)ps_active.size		();
+	Device.Statistic->Particles_destroy	= (u32)ps_destroy.size		();
 
 	// Play req particle systems
 	while (ps_needtoplay.size())
@@ -222,7 +222,7 @@ void IGame_Persistent::destroy_particles		(const bool &all_particles)
 			(*ps_active.begin())->PSI_internal_delete	();
 	}
 	else {
-		u32								active_size = ps_active.size();
+		u32								active_size = (u32)ps_active.size();
 		CPS_Instance					**I = (CPS_Instance**)_alloca(active_size*sizeof(CPS_Instance*));
 		std::copy						(ps_active.begin(),ps_active.end(),I);
